+++ NEWS.md
@@ -2,7 +2,9 @@
 
 -   Add vignette to package
 -   Color adjustment on `bde_rose_pal()`: Now hcl spectrum is more consistent between roses and blues according to `colorspace::specplot()`.
--   Add DOI: https://doi.org/10.5281/zenodo.4673496
+-   Add DOI: <https://doi.org/10.5281/zenodo.4673496>
+-   New palettes for `ggplot2`: `scale_color_bde_c()`, `scale_color_bde_d()`, `scale_fill_bde_c()`, `scale_fill_bde_d()`.
+-   Deprecated scales: `bde_scale_colour_vivid()`, `bde_scale_color_vivid()`, `bde_scale_fill_vivid()`, `bde_scale_colour_rose()`, `bde_scale_color_rose()` `bde_scale_fill_rose()`.
 
 # tidyBdE 0.1.0